+++ class-diagram/.gitignore
@@ -2,3 +2,5 @@
 .vs/
 obj/
 .vscode/
+node_modules/
+package-lock
