--- class-diagram/m-style.css
@@ -1,155 +1,0 @@
-body,
-.dr_lst-wrap,
-.mformedit,
-#dnb-viewcode {
-    scrollbar-width: thin;
-}
-#dnb-vwtab-left {display: inline-flex;flex-direction: column;row-gap:8px;}
-.ctlmove,
-.ctledit {
-    width: 22px;
-    height: calc(100% - 1px);
-    position: absolute;
-    top: 1px;
-    background-color: white;
-    display: none;
-}
-.wrprect {display: inline-block;border: 1px solid rgb(0, 0, 0);
-            position: inherit;cursor: default;border-radius: 4px;}
-.vwheader {
-    font-size: 16px;
-    text-align: center;
-    border-bottom: 1px solid black;
-    padding-left: 24px;
-    padding-right: 45px;
-    position: relative;
-}
-
-.vwheader:hover>.ctlmove,
-.vwheader:hover>.ctledit {
-    display: inline-block;
-}
-
-.cpoint {
-    cursor: pointer;
-}
-
-.p36 {
-    padding: 3px 6px;
-}
-
-.p3-0 {
-    padding: 3px 0;
-}
-
-[contenteditable="true"] {
-    border-bottom: 1px solid #ffffff96;
-}
-
-[contenteditable="true"]:empty:before {
-    content: attr(data-placeholder);
-    color: #98989898;
-    /* Màu placeholder */
-}
-
-.mformedit {
-    min-width: calc(100vw / 3);
-    max-width: calc(100vw - 60px);
-    min-height: 320px;
-    max-height: calc(100vh - 30px);
-    display: inline-block;
-    padding: 9px;
-    transform: translate(-50%, -50%);
-    box-sizing: border-box;
-    border-radius: 10px;
-    position: fixed;
-    overflow-y: hidden;
-    background-color: #e5e5e5;
-    top: 50%;
-    left: 50%;
-}
-
-.fbody {
-    overflow: auto;
-    scrollbar-width: thin;
-    padding-right: 4px;
-    max-height: calc(100vh - 30px - 38px - 36px);
-}
-
-.wrap-mnlist {
-    position: relative;
-    cursor: default;
-    height: 20px;
-}
-
-.dnb-mnlst {
-    display: inline-flex;
-    flex-direction: column;
-    position: absolute;
-    z-index: 1;
-    background-color: white;
-    top: -6px;
-    width: max-content;
-    box-shadow: 0 .125rem .25rem #00000030;
-    border-radius: 6px;
-}
-
-.dnb-mnlst>*:hover {
-    background-color: #efefef;
-}
-
-.dr_lst-wrap>*:hover {
-    background-color: #efefef;
-}
-
-.wrp-drpsearch {
-    display: inline-block;
-    width: 12px;
-    height: 24px;
-    font-size: 12px;
-    font-weight: normal;
-    position: relative;
-}
-
-.drps-wrap {
-    position: absolute; z-index: 1;
-    top: 24px;
-    right: -12px;
-    display: flex;
-    flex-direction: column;
-    min-width: 66px;
-    min-height: 66px;
-    background-color: white;
-    padding: 4px;
-    border-radius: 5px;
-    width: 174px;
-}
-
-.itmdrp {
-    padding: 3px;
-    cursor: pointer;
-    margin-top: 1px;
-    display: grid;
-    grid-template-columns: auto 16px;
-    border-radius: 4px;
-}
-
-.extdrp {
-    position: absolute;
-    top: -9px;
-    right: -9px;
-    display: inline-flex;
-    width: 24px;
-    height: 24px;
-    align-items: center;
-    justify-content: center;
-    background-color: #fcfcfc;
-    border-radius: 6px;
-}
-
-.itmmnls {
-    border-bottom: solid 1px #e3e3e3;
-    box-sizing: border-box;
-    width: 100%;
-    min-width: 105px;
-}
