+++ class-diagram/package.json
@@ -1,0 +1,30 @@
+{
+  "dependencies": {
+    "crypto-js": "^4.2.0",
+    "vue": "^3.5.25"
+  },
+  "name": "class-diagram",
+  "version": "1.0.0",
+  "private": true,
+  "scripts": {
+    "build": "webpack --config prod.config.js",
+    "dev": "webpack serve --mode development"
+  },
+  "keywords": [],
+  "author": "nguyen.ba.dai",
+  "license": "ISC",
+  "description": "",
+  "devDependencies": {
+    "@vue/cli": "^4.2.2",
+    "@vue/compiler-sfc": "^3.5.25",
+    "@webpack-cli/serve": "^3.0.1",
+    "css-loader": "^7.1.2",
+    "html-inline-script-webpack-plugin": "^3.2.1",
+    "html-webpack-plugin": "^5.6.5",
+    "style-loader": "^4.0.0",
+    "vue-loader": "^17.4.2",
+    "webpack": "^5.104.0",
+    "webpack-cli": "^6.0.1",
+    "webpack-dev-server": "^5.2.2"
+  }
+}
