+++ class-diagram/src/Formedit.vue
@@ -1,59 +1,89 @@
+<template>
+    <div class="mformedit" style="z-index: 3;display: grid; grid-template-rows: 38px auto;">
+        <div class="fhead" style="margin-bottom: 12px; display: flex;align-items: center;position: relative;">
+            <menu-list v-if="IsSelectTyp" :value="TxtType" :isfix="true" :sources="StrucTypes"
+                @change:value="$e => changeTypeObject($e)">
+            </menu-list>
+            <span v-else>{{ TxtType }}</span>
+            <span contenteditable="true" @blur="e => onBlurEditable(e, 'class name')"
+                @keypress="e => $root.preventKeyPress(e, [13, 32])" data-placeholder="NameOfClass" class="p36 ceditable"
+                v-html="entry.Name"></span>
+            <i v-for="exv in ViewExtends">{{ exv[0] }}&nbsp;{{ exv[1] }}&nbsp;</i>
+            <drplst-search v-if="IsDrpExtend" :sources="ListExtend" :ids="entry.toIds" :type="entry.TypeDeclaration"
+                @select:id="$e => addExtend($e)" @remove:id="$e => removeExtend($e)"
+                @on:exit="$e => hideDrpSearch()"></drplst-search>
+            <i v-else-if="Extendable" @click.stop="showDrpSearch" class="bi bi-code-square"
+                style="margin-left: 6px;"></i>
+            <i style="width: 24px;height: 20px;display: inline-block;text-align: center;
+          background-color: white;position: absolute;right: 30px;top: -4px;
+          border-radius: 6px;padding-top: 4px;" class="bi bi-floppy cpoint" @click.stop="onSaveChange"></i>
+            <i style="width: 24px;height: 20px;display: inline-block;text-align: center;
+        background-color: white;position: absolute;right: -4px;top: -4px;border-radius: 6px;padding-top: 4px;"
+                class="bi bi-x-lg cpoint" @click.stop="onCloseEdit"></i>
+        </div>
+        <div class="fbody">
+            <div v-if="Extendable != 2">
+                <div v-for="(field, ii) in entry.Fields" :key="'oldfield_' + field.AccessModify"
+                    style="margin-bottom: 2px; display: flex;align-items: center;">
+                    <span v-if="HasPropers" @keypress="e => $root.preventKeyPress(e, [13])"
+                        @blur="e => onBlurEditable(e, 'field acmodify', ii)" data-placeholder="AccessoriOfField"
+                        contenteditable="true" class="p36 ceditable">{{ field.AccessModify }}</span>
+                    <span @keypress="e => $root.preventKeyPress(e, [13, 32])"
+                        @blur="e => onBlurEditable(e, 'field name', ii)" data-placeholder="NameOfField"
+                        class="p36 ceditable" contenteditable="true">{{ field.Name }}</span>
+                    <span v-if="HasPropers" @keypress="e => $root.preventKeyPress(e, [13])"
+                        @blur="e => onBlurEditable(e, 'field type', ii)" data-placeholder="TypeOfField"
+                        class="p36 ceditable" contenteditable="true">{{ field.DataType }}</span>
+                    <span style="margin-left: 12px;cursor: pointer;" @click.stop="e => removeField(ii)"><i
+                            class="bi bi-trash3"></i></span>
+                </div>
+                <span v-if="2 != Extendable" style="cursor: pointer;display: inline-block;" @click.stop="addField">[new
+                    Field]</span>
+            </div>
+            <div v-for="(prp, ii) in entry.Properties" :key="ii" style="margin-bottom: 2px;"
+                :style="[0 == ii ? { marginTop: '10px' } : null]">
+                <div v-if="HasPropers" style="display: flex;align-items: center;">
+                    <span class="p36 ceditable" @keypress="e => $root.preventKeyPress(e, [13])"
+                        @blur="e => onBlurEditable(e, 'methd access type', ii)" data-placeholder="MethodAccessType"
+                        contenteditable="true">{{ prp.AccessModify }}</span>
+                    <menu-list :value="prp.specialMe" :isfix="true" :sources="AccessInit.map(x => x[$root.PLang])"
+                        @change:value="$event => changeAccessor(ii, $event)">
+                    </menu-list>
+                    <span @keypress="e => $root.preventKeyPress(e, [13, 32])"
+                        @blur="e => onBlurEditable(e, 'methd name', ii)" class="p36 ceditable" contenteditable="true"
+                        data-placeholder="NameOfMethod" v-html="prp.Name"></span>
+                    (<span @keypress="e => $root.preventKeyPress(e, [13])"
+                        @blur="e => onBlurEditable(e, 'methd params', ii)" class="p3-0 ceditable" contenteditable="true"
+                        data-placeholder="type name, ...">{{prp.params.map(x => x[1] + ' ' + x[0]).join(', ')}}</span>)
+                    <span v-if="isReturnType(prp.specialMe)" style="padding-left: 3px;"><i>return</i>
+                        <span @blur="e => onBlurEditable(e, 'methd return type', ii)"
+                            @keypress="e => $root.preventKeyPress(e, [13, 32])" data-placeholder="return type"
+                            contenteditable="true" class="p36 ceditable">{{ prp.DataType }}</span>
+                    </span>
+                    <span style="margin-left: 12px;cursor: pointer;" @click.stop="e => removeProperty(ii)"><i
+                            class="bi bi-trash3"></i></span>
+                </div>
+                <textarea v-if="HasCode" class="objedit-vwcode"
+                    style="width: 100%; min-height: 90px;border-radius: 6px;border: none;"
+                    placeholder="ContentCodeOfMethod" @paste="e => pasteTxtArea(e, ii)" @keyup.enter="keyUpTxtArea"
+                    @keyup.delete="keyUpTxtArea" @blur="e => onBlurEditable(e, 'methd body code', ii)"
+                    @change="e => onInput(e, 'methd body code', ii)">{{ prp.FuncBody }}</textarea>
+            </div>
+            <div v-if="HasPropers" style="cursor: pointer;display: inline-block;" @click.stop="addProperty">[new
+                Property]
+            </div>
+        </div>
+    </div>
+</template>
+<script>
 import {
     StructTypes, AccessInit, setHeight, objNewCls,
-    clearSpace, verifySave, convertAccessors,
-    PropName, hasnMethod, removeExtraSpaces,
-} from "../common.js"
-import { MenuList, isAbstract, isInterface, isClass, isEnum, isStruct, } from "./vw-diagram.js"
-export const PopDropdownSearch = {
-    template: `#tmp-drp-search`,
-    name: "Drop_Search",
-    display: "Drop.Search",
-    props: ['ids', 'sources', 'type'],
-    emits: ['select:id', 'remove:id', 'on:exit'],
-    data() {
-        return {
-            TxtSearch: '',
-            ListSrc: this.sources
-        }
-    },
-    mounted() { document.addEventListener('keyup', this.pressEscKey) },
-    methods: {
-        bgSelect(item) {
-            if (this.ids.includes(item.id)) return {
-                backgroundColor: '#d1d1d1'
-            }
-        },
-        inSearch(e) {
-            const target = e.target
-            let str = target.value
-            str = str.trim()
-            this.ListSrc = this.getLstSrc(str)
-        },
-        getLstSrc(str, sources) {
-            if (!sources) sources = this.sources
-            if (typeof str != 'string') return sources
-            if (!str.length) return sources
-            const lst = []
-            for (let ii = 0, src, name; ii < this.sources.length; ii++) {
-                src = this.sources[ii]
-                name = src.Name
-                name = name.toLowerCase()
-                if (name.includes(str)) lst.push(src)
-            }
-            return lst
-        },
-        selectItem(item) { this.$emit('select:id', item.id) },
-        removeItem(item) { this.$emit('remove:id', item.id) },
-        pressEscKey(event) { if ('Escape' == event.key) { this.$emit('on:exit') } },
-    },
-    beforeUnmount() { document.removeEventListener('keyup', this.pressEscKey) },
-    watch: {
-        sources(vals) { this.ListSrc = this.getLstSrc(this.TxtSearch, vals) },
-    },
-}
-
-export const FormEdit = {
-    template: `#tmp-form-edit`,
+    clearSpace, convertAccessors, PropName, hasnMethod, removeExtraSpaces,
+} from "./common.js";
+import { isAbstract, isInterface, isClass, isEnum, isStruct, filterItems, mapItems } from './Appmixin.js';
+import MenuList from './Menulist.vue';
+import PopDropdownSearch from './Dropdownsearch.vue';
+export default {
     name: "Form_Edit",
     display: "Form.Edit",
     components: {
@@ -72,6 +102,7 @@ export const FormEdit = {
             const mItem = this.entry
             if (typeof mItem.id != 'number') return true
             if (isClass(mItem.TypeDeclaration)) return true
+            if (isAbstract(mItem.TypeDeclaration)) return true
             return false
         },
         TxtType() {
@@ -85,11 +116,13 @@ export const FormEdit = {
         HasCode() {
             const type = this.entry.TypeDeclaration
             if (isClass(type)) return true
+            if (isAbstract(type)) return true
             return false
         },
         HasPropers() {
             const type = this.entry.TypeDeclaration
-            if (isClass(type)) return true      // class/abstrac
+            if (isClass(type)) return true
+            if (isAbstract(type)) return true
             if (isInterface(type)) return true
             if (isStruct(type)) return true
             return false
@@ -106,14 +139,12 @@ export const FormEdit = {
             const root = this.$root
             const tIds = this.entry.toIds
             if (!tIds || !tIds.length) return ''
-            let lst = root.ListClass
             const itemId = this.entry.id
             const lsCls = []
             const lstItf = []
-            for (let ii = 0, cls; ii < lst.length; ii++) {
-                cls = lst[ii]
-                if (itemId == cls.id) continue;  // it-self
-                if (!tIds.includes(cls.id)) continue;
+            for (let [id, cls] of root.MpClass) {
+                if (itemId == id) continue;  // it-self
+                if (!tIds.includes(id)) continue;
                 if (isInterface(cls.TypeDeclaration)) lstItf.push(cls)
                 else lsCls.push(cls)
             }
@@ -132,25 +163,23 @@ export const FormEdit = {
             if (!idsTo) return []
             let lst = []
             const itemId = mItem.id
-            let lstSrc = this.$root.ListClass
-            lstSrc = lstSrc.filter(src => src.id != itemId)
-            lstSrc = lstSrc.filter(src => !isEnum(src.TypeDeclaration))
-            lstSrc = lstSrc.filter(src => !isStruct(src.TypeDeclaration))
-            lstSrc = lstSrc.filter(src => !src.toIds || !src.toIds.includes(itemId))
-            let lsEx = []
-            for (let ii = 0, src; ii < lstSrc.length; ii++) {
-                src = lstSrc[ii]
+            let lstSrc = this.$root.MpClass
+            lstSrc = filterItems(lstSrc, (src) => src.id != itemId && !isEnum(src.TypeDeclaration) && !isStruct(src.TypeDeclaration), 'map')
+            lstSrc = filterItems(lstSrc, (src) => !src.toIds || !src.toIds.includes(itemId), 'set')
+
+            let lsEx = new Set()
+            for (let src of lstSrc) {
                 if (idsTo.includes(src.id) && isClass(src.TypeDeclaration))
-                    lsEx.push(src.id)
+                    lsEx.add(src.id)
                 lst.push(src)
             }
             if (isAbstract(mItem.TypeDeclaration)) {
-                if (lsEx.length) {
+                if (lsEx.size) {
                     for (let ii = lst.length - 1, src; -1 < ii; ii--) {
                         src = lst[ii]
                         if (isInterface(src.TypeDeclaration)) continue
-                        if (isAbstract(src.TypeDeclaration)) {
-                            if (lsEx.includes(src.id)) continue
+                        if (isAbstract(src.TypeDeclaration) && lsEx.has(src.id)) {
+                            continue
                         }
                         lst.splice(ii, 1)
                     }
@@ -164,10 +193,10 @@ export const FormEdit = {
                 return lst
             }
             if (isClass(mItem.TypeDeclaration)) {
-                if (lsEx.length) {
+                if (lsEx.size) {
                     for (let ii = lst.length - 1, src; -1 < ii; ii--) {
                         src = lst[ii]
-                        if (lsEx.includes(src.id)) continue
+                        if (lsEx.has(src.id)) continue
                         if (isClass(src.TypeDeclaration)) lst.splice(ii, 1)
                     }
                 }
@@ -184,18 +213,18 @@ export const FormEdit = {
         },
     },
     methods: {
-        changeTypeObject(val) { //console.log('change type object ', val)
+        changeTypeObject(val) {
             const frmCode = this.$root.DynamicVar.get('FrameCode')
             const mItem = frmCode.cItem
             const ii = this.$root.PLang
             switch (val) {
-                case StructTypes[0][ii]: mItem.type = StructTypes[0][0]
+                case StructTypes[0][ii]: mItem.TypeDeclaration = StructTypes[0][0]
                     break;
-                case StructTypes[1][ii]: mItem.type = StructTypes[1][0]
+                case StructTypes[1][ii]: mItem.TypeDeclaration = StructTypes[1][0]
                     break;
-                case StructTypes[2][ii]: mItem.type = StructTypes[2][0]
+                case StructTypes[2][ii]: mItem.TypeDeclaration = StructTypes[2][0]
                     break;
-                case StructTypes[3][ii]: mItem.type = StructTypes[3][0]
+                case StructTypes[3][ii]: mItem.TypeDeclaration = StructTypes[3][0]
                     break;
                 default: return;
             }
@@ -245,9 +274,11 @@ export const FormEdit = {
                     prpF.Name = txtC
                     break;
                 case 'methd params':
-                    prpF = mItem.Properties[ii]; debugger
-                    if (checkRevertHtml(prpF[2])) return
-                    prpF[2] = removeExtraSpaces(txtC)
+                    prpF = mItem.Properties[ii]; 
+                    let params = txtC.split(',').filter(x => !!x)
+                    params = params.map(x => removeExtraSpaces(x))
+                    params = params.map(x => x.split(' ').filter(t => !!t))
+                    prpF.params = params.map(x => [x[1], x[0]])
                     break;
                 case 'methd return type':
                     prpF = mItem.Properties[ii]
@@ -343,14 +374,9 @@ export const FormEdit = {
         addProperty() {
             const frmCode = this.$root.DynamicVar.get('FrameCode')
             const mItem = frmCode.cItem
-            if (hasnMethod(mItem)) {
-                mItem.Properties = [
-                    ['public', PropName, '', 'void', AccessInit[1][0], '']
-                ]
-            } else {
-                if (mItem.Properties.find(x => PropName == x[1])) return;
-                mItem.Properties.push(['public', PropName, '', 'void', AccessInit[1][0], ''])
-            }
+            if (!mItem.Properties) mItem.Properties = []
+            if (mItem.Properties.find(x => PropName == x.Name)) return;
+            mItem.Properties.push({ Name: '', params: [['a', 'int']], DataType: 'void', FuncBody: '', specialMe: 'set', AccessModify: 'public' },)
         },
         onCloseEdit() {
             const root = this.$root
@@ -361,7 +387,7 @@ export const FormEdit = {
             const root = this.$root
             const frmCode = root.DynamicVar.get('FrameCode')
             const mItem = frmCode.cItem
-            const lstCls = root.ListClass;
+            let mapCls = root.MpClass;
             if (typeof mItem.id != 'number') {
                 let nItem = onNewItem.call(this, mItem)
                 root.NewClassName = null
@@ -374,10 +400,11 @@ export const FormEdit = {
                         root.$nextTick(root.drawInCnvs) // new item
                     }
                 }
+                root.MpClass = mapCls
                 return
             }
             const item = frmCode.item
-            item.type = mItem.type
+            item.TypeDeclaration = mItem.TypeDeclaration
             let name = mItem.Name
             name = clearSpace(name, item.Name)
             if (!name.length) {
@@ -387,20 +414,18 @@ export const FormEdit = {
             root.MpPoints.clear()
             item.Name = name
             item.toIds = mItem.toIds
-            verifySave(mItem, root.PLang, true)
-
+            
             item.Fields = mItem.Fields
             item.Properties = mItem.Properties
+            for (let [id, cls] of mapCls) root.buildMapPoints(cls)      // save change
 
-            for (let ii = lstCls.length - 1; -1 < ii; ii--) {
-                root.buildMapPoints(lstCls[ii])      // save change
-            }
             root.buildAssociation()             // save change
             this.onCloseEdit()
             root.$nextTick(root.updateSizeCanvas) // save change
             root.$nextTick(root.drawInCnvs) // save change
 
             function onNewItem(newItem) {
+                const maxId = Math.max(...mapItems(mapCls, x => x.id))
                 let nItem = objNewCls(newItem)
                 if (!nItem) {
                     root.NewClassName = null
@@ -411,15 +436,15 @@ export const FormEdit = {
                     return;
                 }
                 nItem = verifyNewItem.call(this, nItem)
-                verifySave(nItem, root.PLang, true)
-                if (!lstCls.length) nItem.id = 1
-                lstCls.push(nItem)
+                if (!mapCls.size) nItem.id = 1
+                mapCls.set(nItem.id, nItem)
+                mapCls = new Map(mapCls)    // new ref
                 return nItem
                 function verifyNewItem(item) {
                     let newName = item.Name
                     if (!newName.length) return item
-                    const maxId = Math.max(...lstCls.map(x => x.id))
-                    let lstNo = lstCls.map(x => x.Name)
+
+                    let lstNo = mapItems(mapCls, x => x.Name)
                     if (lstNo.includes(newName)) {
                         let index = 1
                         newName = `${item.Name}${index}`
@@ -459,3 +484,26 @@ export const FormEdit = {
         this.$root.DynamicVar.delete('Drop-Search')
     },
 }
+</script>
+<style scoped>
+.mformedit {
+    scrollbar-width: thin;
+}
+
+.mformedit {
+    min-width: calc(100vw / 3);
+    max-width: calc(100vw - 60px);
+    min-height: 320px;
+    max-height: calc(100vh - 30px);
+    display: inline-block;
+    padding: 9px;
+    transform: translate(-50%, -50%);
+    box-sizing: border-box;
+    border-radius: 10px;
+    position: fixed;
+    overflow-y: hidden;
+    background-color: #e5e5e5;
+    top: 50%;
+    left: 50%;
+}
+</style>
