+++ class-diagram/src/main.js
@@ -1,0 +1,5 @@
+import { createApp } from 'vue';
+import App from './App.vue';
+createApp(App).mount('#dnbApp');
+// import Demo from './Demo.vue';
+// createApp(Demo).mount('#dnbApp');
